requirements.txt
streamlit>=1.32.0,<2.0.0
pandas>=2.2.0,<3.0.0
plotly>=5.20.0,<6.0.0
groq>=0.4.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
